changed custom view height into %. added BaseCustomView for custom views
