+++ .idea/misc.xml
@@ -11,7 +11,7 @@
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/activity_main.xml" value="0.17552083333333332" />
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/fragment_home.xml" value="0.22" />
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/fragment_questions.xml" value="0.17552083333333332" />
-        <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/fragment_start.xml" value="0.13333333333333333" />
+        <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/fragment_start.xml" value="0.16666666666666666" />
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/fragment_start_screen.xml" value="0.17552083333333332" />
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/home_screen.xml" value="0.17447916666666666" />
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/questions_screen.xml" value="0.131640625" />
