+++ app/src/main/java/com/space/quiz_app/presentation/quiz_activity/QuizActivity.kt
@@ -10,6 +10,7 @@ class QuizActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+        supportActionBar?.hide()
         setUpNavigation()
     }
