+++ app/src/main/java/com/space/quiz_app/presentation/quiz_start_screen/StartScreenFragment.kt
@@ -13,7 +13,7 @@ class StartScreenFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_home, container, false)
+        return inflater.inflate(R.layout.fragment_start, container, false)
     }
