+++ app/src/main/res/layout/fragment_start.xml
@@ -9,7 +9,8 @@
 
     <com.space.quiz_app.presentation.custom_view.CustomStartBackground
         android:layout_width="match_parent"
-        android:layout_height="430dp"
+        android:layout_height="0dp"
+        app:layout_constraintHeight_percent="0.6"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
@@ -22,6 +23,7 @@
         android:layout_marginTop="@dimen/dimen_80_dp"
         android:layout_marginBottom="@dimen/dimen_80_dp"
         android:text="@string/welcome_text"
+        android:textColor="@color/white"
         app:layout_constraintBottom_toTopOf="@+id/iconImageView"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -43,9 +45,9 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:minHeight="@dimen/dimen_50_dp"
-        android:layout_gravity="center"
-        android:layout_marginHorizontal="@dimen/dimen_80_dp"
-        android:layout_marginBottom="@dimen/dimen_20_dp"
+        android:textAlignment="center"
+        android:layout_marginHorizontal="@dimen/dimen_60_dp"
+        android:layout_marginBottom="@dimen/dimen_30_dp"
         android:background="@drawable/shp_message_edittext"
         android:hint="@string/hint_text"
         android:paddingStart="@dimen/dimen_30_dp"
